% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImportData.R
\name{attachModal}
\alias{attachModal}
\title{attachModal (not finished)}
\usage{
attachModal(object = NULL, data = NULL, target.col.name = NULL,
  type = c("bool", "value"), barcode.index = 1)
}
\arguments{
\item{object}{Seurat object}

\item{data}{data table contains modality information with barcode}

\item{target.col.name}{ssss}

\item{type}{could be "bool" or "value"}

\item{barcode.index}{index of barcode column in data table}
}
\description{
read modalities and attach to cells
}
