% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildGraphTree.R
\name{batchPseudoTime}
\alias{batchPseudoTime}
\title{batchPseudoTime}
\usage{
batchPseudoTime(object = NULL, method = c("MST", "KNN"),
  mst = "Joint", knn = "Joint", dist = NULL, root = NULL)
}
\arguments{
\item{object}{Seurat object}

\item{method}{could be MST or KNN}

\item{mst}{user choose the assay data MST they want to use for pseudotime, can be RNA, ADT, Joint or All (All means calculate all three), use also can determine any mst (for example, mst = "rnaTree", the rnaTree must be located in object[at]misc[['rnaTree']]}

\item{knn}{user choose the assay data KNN they want to use for pseudotime, can be RNA, ADT, Joint or All (All means calculate all three), use also can determine any knn by name (for example, knn = "rnaGraph", the rnaTree must be located in object[at]misc[['rnaGraph']]}

\item{dist}{if use choose their own KNN or MST, user should determine the dist object or distance matrix for the knn graph or MST. the dist can be dist object, or the name of the dist object that is located in object[at]misc[[dist]]. for example, user can set dist = object[at]misc[["userDist"]] or knn.dist = "userDist"}

\item{root}{user choose the root cell}
}
\description{
calculate pseudotime using a given start point from MST or KNN. This function only can be used if you know which cell is the root. Not suggested.
}
