% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildGraphTree.R, R/generics.R
\name{buildKNN.Seurat}
\alias{buildKNN.Seurat}
\alias{buildKNN.default}
\alias{buildKNN}
\title{buildKNN.Seurat}
\usage{
\method{buildKNN}{Seurat}(object, assay = "Joint", k = 10)

\method{buildKNN}{default}(object, assay, k = 10)

buildKNN(object, ...)
}
\arguments{
\item{object}{Seurat object, or dist object}

\item{assay}{(For Seurat) user choose the assay data they want to build a MST based on, can be RNA, ADT, Joint or All (All means calculate all three)}

\item{k}{k value for KNN graph. Default = 10}
}
\description{
build knn network based on user's choice (RNA, ADT, Joint or All)
}
