% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildGraphTree.R, R/generics.R
\name{buildMST.Seurat}
\alias{buildMST.Seurat}
\alias{buildMST.dist}
\alias{buildMST.matrix}
\alias{buildMST.default}
\alias{buildMST}
\title{buildMST.Seurat}
\usage{
\method{buildMST}{Seurat}(object, assay = "Joint")

\method{buildMST}{dist}(object)

\method{buildMST}{matrix}(object)

\method{buildMST}{default}(object)

buildMST(object, ...)
}
\arguments{
\item{object}{Seurat object, or dist object, or N*N distance martix}

\item{assay}{If object = Seurat object, user choose the assay data they want to build a MST based on, can be RNA, ADT, Joint or All (All means calculate all three)}
}
\description{
If input Seurat object, build Minimum spanning tree (MST) based on user's choice (RNA, ADT, Joint or All). Also can build MST for any input dist object
}
