% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildGraphTree.R, R/generics.R
\name{buildPG.Seurat}
\alias{buildPG.Seurat}
\alias{buildPG.default}
\alias{buildPG}
\title{buildPG.Seurat}
\usage{
\method{buildPG}{Seurat}(object, assay = "Joint",
  reduction.prefix = "tsne_", pg.nodes = NULL, pg.min.nodes = 30,
  pg.Lambda = 0.03, pg.Mu = 0.01, trimming.radius = Inf,
  final.energy = "Penalized", initial.MST = FALSE)

\method{buildPG}{default}(object, mst = NULL, pg.nodes = NULL,
  pg.min.nodes = 30, pg.Lambda = 0.03, pg.Mu = 0.01,
  trimming.radius = Inf, final.energy = "Penalized",
  initial.MST = FALSE)

buildPG(object, ...)
}
\arguments{
\item{object}{(For Seurat) Seurat object}

\item{assay}{(For Seurat) user choose the assay data they want to build a MST based on, can be RNA, ADT, Joint or All (All means calculate all three)}

\item{reduction.prefix}{(For Seurat) the prefix of reduction. e.g. "tsne_" or "umap_"}

\item{pg.nodes}{parameter for computing Elastic Principal Graph. number of nodes you want to build PG. if NULL, we will calculated a number based on your data size}

\item{pg.min.nodes}{parameter for computing Elastic Principal Graph. min value of pg.nodes}

\item{pg.Lambda}{parameter for computing Elastic Principal Graph, real, the lambda parameter used the compute the elastic energy}

\item{pg.Mu}{parameter for computing Elastic Principal Graph, real, the lambda parameter used the compute the elastic energy}

\item{trimming.radius}{parameter for computing Elastic Principal Graph, real, maximal distance of point from a node to affect its embedment}

\item{final.energy}{parameter for computing Elastic Principal Graph, string indicating the final elastic emergy associated with the configuration. Currently it can be "Base" or "Penalized"}

\item{initial.MST}{use MST as initial for PG or not. default is FALSE because use MST as initial is time consuming for large dataset}

\item{mst}{(For default) MST for building PG}

\item{object}{(For default) cell embedding}
}
\description{
build PG based on user's choice (RNA, ADT, Joint or All) MST
}
