% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization.R
\name{clusterConnectionPlot}
\alias{clusterConnectionPlot}
\title{clusterConnectionPlot}
\usage{
clusterConnectionPlot(object = NULL, group.by = NULL,
  reduction = NULL, graph = NULL, cutoff = 0.01,
  line.color = "black", point.size = 2, path.size = 1.5)
}
\arguments{
\item{object}{seurat object}

\item{group.by}{cell cluster name}

\item{reduction}{reduction name}

\item{graph}{a KNN graph name}

\item{cutoff}{only connections between two cluster have connection factor larger than this cutoff will be shown}

\item{line.color}{line color of the connections between cell clusters}

\item{point.size}{control dot size of cell clusters}

\item{path.size}{control path size of cell connections}
}
\description{
plot cell clusters with connections
}
