% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization.R
\name{colorGradient}
\alias{colorGradient}
\title{colorGradient}
\usage{
colorGradient(x, colors = c("red", "yellow", "green", "blue"),
  colsteps = 100)
}
\arguments{
\item{x}{value vector}

\item{colors}{color plate}

\item{colsteps}{color steps}
}
\description{
get gradient colors
}
