% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization.R
\name{computeSmoothedStacks}
\alias{computeSmoothedStacks}
\title{The function that implements the ThemeRiver and Zero Value algorithms (curved lines). This function was adopted from https://canonicalized.com/streamgraphs-in-tableau-via-r/, author is Dorian Banutoiu}
\usage{
computeSmoothedStacks(values, method = "ThemeRiver", multiple)
}
\value{
ThemeRiver data or Zero Value data (curved lines)
}
\description{
The function that implements the ThemeRiver and Zero Value algorithms (curved lines). This function was adopted from https://canonicalized.com/streamgraphs-in-tableau-via-r/, author is Dorian Banutoiu
}
