% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization.R
\name{computeStacks}
\alias{computeStacks}
\title{The function that implements the ThemeRiver and Zero Value algorithms. This function was adopted from https://canonicalized.com/streamgraphs-in-tableau-via-r/, author is Dorian Banutoiu}
\usage{
computeStacks(values, method = "ThemeRiver")
}
\value{
ThemeRiver data or Zero Value data
}
\description{
The function that implements the ThemeRiver and Zero Value algorithms. This function was adopted from https://canonicalized.com/streamgraphs-in-tableau-via-r/, author is Dorian Banutoiu
}
