% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImportData.R
\name{createObject}
\alias{createObject}
\title{createObject}
\usage{
createObject(data = NULL, project = "SeuratProject", min.cells = 0,
  min.features = 0, names.field = 1, names.delim = "_",
  meta.data = NULL, min.cells.adt = 0, min.features.adt = 0,
  rna = NULL, adt = NULL)
}
\arguments{
\item{data}{data of 10X cell ranger output}

\item{project}{project name for current dataset}

\item{min.cells}{Same as CreateSeuratObject function in Seurat. Include features detected in at least this many cells. Will subset the counts matrix as well. To reintroduce excluded features, create a new object with a lower cutoff.}

\item{min.features}{Same as CreateSeuratObject function in Seurat. Include cells where at least this many features are detected.}

\item{names.field}{Same as CreateSeuratObject function in Seurat. For the initial identity class for each cell, choose this field from the cell's name. E.g. If your cells are named as BARCODE_CLUSTER_CELLTYPE in the input matrix, set names.field to 3 to set the initial identities to CELLTYPE.}

\item{names.delim}{Same as CreateSeuratObject function in Seurat. For the initial identity class for each cell, choose this delimiter from the cell's column name. E.g. If your cells are named as BARCODE-CLUSTER-CELLTYPE, set this to "-" to separate the cell name into its component parts for picking the relevant field.}

\item{meta.data}{Same as CreateSeuratObject function in Seurat. Additional cell-level metadata to add to the Seurat object. Should be a data frame where the rows are cell names and the columns are additional metadata fields.}

\item{min.cells.adt}{Same as CreateAssayObject function in Seurat. Include ADT features detected in at least this many cells. Will subset the counts matrix as well. To reintroduce excluded features, create a new object with a lower cutoff.}

\item{min.features.adt}{Same as CreateAssayObject function in Seurat. Include cells where at least this many ADT features are detected.}

\item{assay}{assay name, default is RNA}
}
\description{
Create Seurat V3 object from raw data by calling Seurat CreateSeuratObject function and CreateAssayObject function
}
