% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PreProcess.R
\name{dataNormalization}
\alias{dataNormalization}
\title{dataNormalization}
\usage{
dataNormalization(object = NULL, scale.factor = 10000, margin = 1,
  verbose = TRUE)
}
\arguments{
\item{object}{Seurat object}

\item{scale.factor}{Sets the scale factor for cell-level normalization. Default is 10000}

\item{margin}{If performing CLR normalization, normalize across features (1) or cells (2)}

\item{verbose}{display progress bar for normalization procedure}
}
\description{
data normalization for both RNA and ADT data
a key step for pre-processing. Call seurat function
}
