% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PreProcess.R
\name{dataScaling}
\alias{dataScaling}
\title{dataScaling}
\usage{
dataScaling(object = NULL, features = NULL, assay = NULL,
  vars.to.regress = NULL, model.use = "linear", use.umi = FALSE,
  do.scale = TRUE, do.center = TRUE, scale.max = 10,
  block.size = 1000, min.cells.to.block = 3000, verbose = TRUE)
}
\arguments{
\item{object}{= NULL, A Seurat object}

\item{features}{= NULL, features on data scaling}

\item{assay}{= NULL,  assay name, RNA, ADT, HTO}

\item{vars.to.regress}{= NULL, vars.to.regress, same as Seurat,  Variables to regress out (previously latent.vars in RegressOut). For example, nUMI, or percent.mito.}

\item{model.use}{= "linear", same as Seurat, Use a linear model or generalized linear model (poisson, negative binomial) for the regression. Options are 'linear' (default), 'poisson', and 'negbinom'}

\item{use.umi}{= FALSE, same as Seurat, Regress on UMI count data. Default is FALSE for linear modeling, but automatically set to TRUE if model.use is 'negbinom' or 'poisson'}

\item{do.scale}{= TRUE, same as Seurat, Whether to scale the data.}

\item{do.center}{= TRUE, same as Seurat, Whether to center the data.}

\item{scale.max}{= 10,  same as Seurat,  Max value to return for scaled data. The default is 10. Setting this can help reduce the effects of feautres that are only expressed in a very small number of cells.}

\item{block.size}{= 1000, same as Seurat, Default size for number of feautres to scale at in a single computation. Increasing block.size may speed up calculations but at an additional memory cost.}

\item{min.cells.to.block}{= 3000, same as Seurat, If object contains fewer than this number of cells, don't block for scaling calculations.}

\item{verbose}{= TRUE, same as Seurat, Displays a progress bar for scaling procedure}
}
\description{
data scaling for both RNA and ADT (for later Heatmap viosualization)
}
