% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization.R
\name{DimplotContinue}
\alias{DimplotContinue}
\title{DimplotContinue}
\usage{
DimplotContinue(object = NULL, reduction = "pca", color.by = NULL,
  colors = c("blue", "green", "yellow", "red"), display.min = NULL,
  display.max = NULL, outlier.color = "white", xlab = "dim1",
  ylab = "dim2")
}
\arguments{
\item{object}{seurat object}

\item{reduction}{reduction name}

\item{color.by}{a series of continuous values (e.g. percent.mt)}

\item{colors}{gradient colors for pseudotime. color order is low to high.}

\item{display.min}{min value to display, color of excluded cells will be indicated by outlier.color}

\item{display.max}{max value to display}

\item{outlier.color}{color of outlier cells when display.min or display.max was set}

\item{xlab}{label of x axis}

\item{ylab}{label of y axis}
}
\description{
plot cells on t-SNE or UMAP embedding and colored by continuous values (e.g. percent.mt)
}
