% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization.R
\name{distHeatMap}
\alias{distHeatMap}
\title{distHeatMap}
\usage{
distHeatMap(
  object = NULL,
  assay = "Joint",
  cell.num = 1000,
  group.by = NULL,
  colors = c("yellow", "green"),
  group.bar = TRUE,
  label = FALSE,
  size = 5.5,
  hjust = 0,
  angle = 90,
  dims = 20
)
}
\arguments{
\item{object}{seurat object}

\item{assay}{Assay name. Default is "Joint"}

\item{cell.num}{number of cells in heatmap. For large datasets, user may want to reduce the number of cells. Default value is set to 1000, indicate 1000 cells at maximum.}

\item{group.by}{order cells by cell clusters. If set to NULL, will use current ident.}

\item{colors}{two distinct colors for ADT and RNA.}

\item{group.bar}{Add a color bar showing group status for cells}

\item{label}{Label the cell identies above the color bar}

\item{size}{Size of text above color bar}

\item{hjust}{Horizontal justification of text above color bar}

\item{angle}{angle of cluster label}

\item{dims}{If RNA distance is not available, dims of PCA used to calculate RNA distance.}
}
\description{
plot contribution of each modality on a N*N heatmap
}
