% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JointAnalysis.R
\name{errorFunction}
\alias{errorFunction}
\title{errorFunction}
\usage{
errorFunction(alpha, X, Y)
}
\arguments{
\item{alpha}{current alpha}

\item{X}{dist of X}

\item{Y}{dist of Y}
}
\description{
objective function value
}
