% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildGraphTree.R
\name{fastKNN}
\alias{fastKNN}
\title{fastKNN}
\usage{
fastKNN(dist = NULL, k = NULL)
}
\arguments{
\item{dist}{dist object or matrix}

\item{k}{k value for KNN graph}
}
\description{
run fastKNN method
}
