% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildGraphTree.R
\name{findRootNodeID}
\alias{findRootNodeID}
\title{findRootNodeID}
\usage{
findRootNodeID(dist = NULL, root.cluster.index = NULL,
  other.cluster.index = NULL, option = 1)
}
\arguments{
\item{dist}{a MST}

\item{root.cluster.index}{user choose the a root node}
}
\description{
calculate pseduotime using a given start point from MST
}
