% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JointAnalysis.R
\name{gradientDescent}
\alias{gradientDescent}
\title{gradientDescent}
\usage{
gradientDescent(X, Y, alpha, learning.rate, low.threshold)
}
\arguments{
\item{X}{dist of X}

\item{Y}{dist of Y}

\item{alpha}{starting alpha}

\item{learning.rate}{step size of each iteration}

\item{low.threshold}{low threshold of gradient}
}
\description{
gradient descent algorithm
}
