% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gradientDescentCpp}
\alias{gradientDescentCpp}
\title{gradient descnet method}
\usage{
gradientDescentCpp(X, Y, alpha, learning_rate, low_threshold, max_iter)
}
\arguments{
\item{X}{distance vector.}

\item{Y}{distance vector.}

\item{alpha}{initial value of parameter alpha.}

\item{learning_rate}{learning rate of GD method}

\item{low_threshold}{the low threshold of GD}

\item{max_iter}{maximum iterations}
}
\description{
gradient descnet method
}
