% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JointAnalysis.R
\name{gradientFunction}
\alias{gradientFunction}
\title{gradientFunction}
\usage{
gradientFunction(alpha, X, Y)
}
\arguments{
\item{alpha}{current alpha}

\item{X}{dist of X}

\item{Y}{dist of Y}
}
\description{
gradient value
}
