% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gradientFunctionCpp}
\alias{gradientFunctionCpp}
\title{gradient function for gradient descnet method}
\usage{
gradientFunctionCpp(alpha, X, Y)
}
\arguments{
\item{alpha}{current value of parameter alpha.}

\item{X}{distance vector.}

\item{Y}{distance vector.}
}
\description{
gradient function for gradient descnet method
}
