% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization.R
\name{gridDimPlot}
\alias{gridDimPlot}
\title{gridDimPlot}
\usage{
gridDimPlot(object = NULL, assays = c("RNA", "ADT", "Joint"),
  reduction.prefix = "umap_", cluster.suffix = "ClusterID",
  darkTheme = TRUE, legend = TRUE, figure.label.size = 7,
  cluster.label = TRUE, cluster.lable.size = 5, wide.rel = 1.5,
  height.rel = 1)
}
\arguments{
\item{object}{seurat object}

\item{assays}{assays user want to plot}

\item{reduction.prefix}{the prefix for reduction name. e.g. "umap_" for UMAP and "tsne" for "t-SNE"}

\item{cluster.suffix}{the suffix for cell cluster name in mate information.}

\item{darkTheme}{switch for darkTheme (TRUE or FALSE)}

\item{legend}{switch for showing legend in each sub plot (TRUE or FALSE)}

\item{figure.label.size}{font size for figure labels}

\item{cluster.label}{swicth for showing cluster label on each sub plot (TRUE or FALSE)}

\item{cluster.lable.size}{font size for cluster labels}

\item{wide.rel}{relative width of figure label to the left. size of plots is 3.}

\item{height.rel}{relative height of figure label on the top. size of plots is 3.}
}
\description{
grid dim plot for RNA, ADT and joint analysis
}
