% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization.R
\name{heatMapPlot}
\alias{heatMapPlot}
\title{heatMapPlot}
\usage{
heatMapPlot(object = NULL, adt.feature = NULL, rna.feature = NULL,
  group.by = NULL, adt.label = TRUE, rna.label = FALSE,
  label.hjust = 0.1, label.size = 5, height.rel = 1,
  legend = FALSE)
}
\arguments{
\item{object}{seurat object}

\item{adt.feature}{ADT features used for plot}

\item{rna.feature}{RNA features used for plot}

\item{group.by}{cell clusters}

\item{adt.label}{annotate ADT features on HEATmap (TRUE or FALSE)}

\item{rna.label}{annotate RNA features on HEATmap (TRUE or FALSE)}

\item{label.hjust}{hjust for ADT labels, number between 0 and 1}

\item{label.size}{font size of ADT labels}

\item{height.rel}{reletive height of RNA compare to ADT. Default 1 means equal height. set to null , the height will be set according to number of features}

\item{legend}{show legend or not. TRUE or FALSE}
}
\description{
plot heatmap of ADT and selected RNA features of cell clusters
}
