% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization.R
\name{highCorrelatedGenePlot}
\alias{highCorrelatedGenePlot}
\title{highCorrelatedGenePlot}
\usage{
highCorrelatedGenePlot(object = NULL, target.gene = NULL,
  coef = "spearman", genes = NULL, topx = 20, group.by = "ident")
}
\arguments{
\item{object}{Seurat object}

\item{target.gene}{target gene user want focus on}

\item{coef}{coef methods: spearman, pearson or kendall}

\item{genes}{user's choice gene list}

\item{topx}{number of genes that highly correlated with target gene}
}
\description{
order all cells by the gene expression level of a specific gene, and explorer the gene expression profiles for highly corrolated genes or a group of user choosed genes
}
