% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JointAnalysis.R
\name{jointClusteing}
\alias{jointClusteing}
\title{jointClusteing
1) Calculate cell-cell pairwise distances for RNA and ADT separately, then calculate the joint cell-cell pairwise distances
2) Run UMAP to project single cells into 2D map using cell-cell pairwise distances
3) Run clustering method (call Seurat function) to identify cell populations using cell-cell pairwise distances
Since trajectory analysis require distances, this function is not recommended}
\usage{
jointClusteing(object, run_seprate = FALSE, reduction_method = "UMAP",
  dims = 20, resolution = c(0.2, 0.2, 0.2), alpha = 0.5)
}
\arguments{
\item{object}{Seurat object}

\item{run_seprate}{also calculate cell-cell pairwise distances, umap and clustering for RNA and ADT separately}

\item{reduction_method}{reduction method. Now only support UMAP. Will support t-SNE later...}

\item{dims}{number of PCs used for RNA data. Default is top 20}

\item{resolution}{resolution for 1) Joint, 2) RNA and 3) ADT clustering. if run_seprate = TRUE, user should provide all three solutions, otherwise will use default value 0.2 for all three datasets}

\item{alpha}{use alpha to balence contributions from RNA and ADT in the joint distances. We suggest use 0.5 for initial analysis, and then adjust alpha for better results}
}
\description{
jointClusteing
1) Calculate cell-cell pairwise distances for RNA and ADT separately, then calculate the joint cell-cell pairwise distances
2) Run UMAP to project single cells into 2D map using cell-cell pairwise distances
3) Run clustering method (call Seurat function) to identify cell populations using cell-cell pairwise distances
Since trajectory analysis require distances, this function is not recommended
}
