% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JointAnalysis.R, R/generics.R
\name{jointDistance.Seurat}
\alias{jointDistance.Seurat}
\alias{jointDistance.default}
\alias{jointDistance}
\title{jointDistance.Seurat}
\usage{
\method{jointDistance}{Seurat}(object, dims = 20, alpha = NULL)

\method{jointDistance}{default}(object, dist1 = NULL, dist2 = NULL,
  alpha = NULL)

jointDistance(object, ...)
}
\arguments{
\item{object}{(For Seurat) Seurat object}

\item{dims}{(For Seurat) number of PCs used for RNA data. Default is 20}

\item{alpha}{(For Seurat) use alpha to balence contributions from RNA and ADT in the joint distances. We suggest use 0.5 for initial analysis, and then adjust alpha for better results. User can set alpha = null, then we will automatically determine the value}

\item{dist1}{(For default) dist object of first assay}

\item{dist2}{(For default) dist object of second assay}

\item{alpha}{(For default) use alpha to balence contributions from two assays in joint distances. We suggest use 0.5 for initial analysis, and then adjust alpha for better results. User can set alpha = null, then we will automatically determine the value}
}
\description{
If input seurat object, will calculate cell-cell pairwise distances for RNA and ADT separately, then calculate the joint cell-cell pairwise distances. It alao can calculate joint distances for any two dist objects
}
