% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization.R
\name{listPlot}
\alias{listPlot}
\title{listPlot}
\usage{
listPlot(
  object = NULL,
  align = "none",
  axis = "none",
  fig.id = NULL,
  fig.ident = NULL,
  fig.map = NULL,
  ncol = NULL,
  rel_widths = 1,
  rel_heights = 1,
  labels = NULL,
  label_size = 10,
  label_fontfamily = NULL,
  label_fontface = "bold",
  label_colour = NULL,
  label_x = 0.1,
  label_y = 1,
  hjust = -0.5,
  vjust = 1,
  scale = 1
)
}
\arguments{
\item{object}{list of dim plots}

\item{align}{parameter of plot_grid() from cowplot package.(optional) Specifies whether graphs in the grid should be horizontally ("h") or vertically ("v") aligned. Options are "none" (default), "hv" (align in both directions), "h", and "v".}

\item{axis}{parameter of plot_grid() from cowplot package.(optional) Specifies whether graphs should be aligned by the left ("l"), right ("r"), top ("t"), or bottom ("b") margins. Options are "none" (default), or a string of any combination of l, r, t, and b in any order (e.g. "tblr" or "rlbt" for aligning all margins). Must be specified if any of the graphs are complex (e.g. faceted) and alignment is specified and desired. See align_plots() for details.}

\item{fig.id}{Numerical ids for selected plots. could be number or vector}

\item{fig.ident}{ident property of selected plots. For example, set to "RNA" will select and plot all plots that have RNA clusters}

\item{fig.map}{map property of selected plots. For example, set to "RNA" will select and plot all plots that have RNA maps}

\item{ncol}{parameter of plot_grid() from cowplot package.(optional) Number of columns in the plot grid.}

\item{rel_widths}{parameter of plot_grid() from cowplot package. (optional) Numerical vector of relative columns widths. For example, in a two-column grid, rel_widths = c(2, 1) would make the first column twice as wide as the second column.}

\item{rel_heights}{parameter of plot_grid() from cowplot package.(optional) Numerical vector of relative columns heights. Works just as rel_widths does, but for rows rather than columns.}

\item{labels}{parameter of plot_grid() from cowplot package.(optional) List of labels to be added to the plots. You can also set labels="AUTO" to auto-generate upper-case labels or labels="auto" to auto-generate lower-case labels.}

\item{label_size}{parameter of plot_grid() from cowplot package.(optional) Numerical value indicating the label size. Default is 14.}

\item{label_fontfamily}{parameter of plot_grid() from cowplot package.(optional) Font family of the plot labels. If not provided, is taken from the current theme.}

\item{label_fontface}{parameter of plot_grid() from cowplot package.(optional) Font face of the plot labels. Default is "bold".}

\item{label_colour}{parameter of plot_grid() from cowplot package.(optional) Color of the plot labels. If not provided, is taken from the current theme.}

\item{label_x}{parameter of plot_grid() from cowplot package.(optional) Single value or vector of x positions for plot labels, relative to each subplot. Defaults to 0 for all labels. (Each label is placed all the way to the left of each plot.)}

\item{label_y}{parameter of plot_grid() from cowplot package.(optional) Single value or vector of y positions for plot labels, relative to each subplot. Defaults to 1 for all labels. (Each label is placed all the way to the top of each plot.)}

\item{hjust}{parameter of plot_grid() from cowplot package.Adjusts the horizontal position of each label. More negative values move the label further to the right on the plot canvas. Can be a single value (applied to all labels) or a vector of values (one for each label). Default is -0.5.}

\item{vjust}{parameter of plot_grid() from cowplot package.Adjusts the vertical position of each label. More positive values move the label further down on the plot canvas. Can be a single value (applied to all labels) or a vector of values (one for each label). Default is 1.5.}

\item{scale}{parameter of plot_grid() from cowplot package.Individual number or vector of numbers greater than 0. Enables you to scale the size of all or select plots. Usually it's preferable to set margins instead of using scale, but scale can sometimes be more powerful.}
}
\description{
plot grid dim plots. the object should be generated by generateGridDimPlot() function
}
