% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImportData.R
\name{loadADT}
\alias{loadADT}
\title{loadADT}
\usage{
loadADT(object = NULL, file = NULL)
}
\arguments{
\item{object}{Seurat object}

\item{file}{hash tag csv file}
}
\description{
load ADT information to a Seurat object from a CSV or TXT file
}
