% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImportData.R
\name{loadHashTag}
\alias{loadHashTag}
\title{loadHashTag}
\usage{
loadHashTag(object = NULL, file = NULL)
}
\arguments{
\item{object}{Seurat object}

\item{file}{hash tag csv file}
}
\description{
load cell hash tag information to a Seurat object from a CSV or TXT file
}
