% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImportData.R
\name{loadSgRNA}
\alias{loadSgRNA}
\title{loadSgRNA}
\usage{
loadSgRNA(object = NULL, file = NULL)
}
\arguments{
\item{object}{Seurat object}

\item{file}{hash tag csv file}
}
\description{
load SgRNA information to a Seurat object from a CSV or TXT file
}
