% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImportData.R
\name{loadTable}
\alias{loadTable}
\title{loadTable}
\usage{
loadTable(file = NULL, header = header)
}
\arguments{
\item{file}{csv file}

\item{header}{the file has a header or not}
}
\description{
load TCR/BCR information from a CSV file to a table
}
