% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImportData.R
\name{loadTCR}
\alias{loadTCR}
\title{loadTCR (not finished)}
\usage{
loadTCR(object = NULL, file = NULL, table = NULL, field = NULL)
}
\arguments{
\item{object}{Seurat object}

\item{file}{TCR/BCR information file path}

\item{table}{TCR/BCR information table}

\item{field}{hash tag csv file}
}
\description{
load TCR/BCR information from a table to a seurat object. Only information in user selected fields will be loaded into seurat object.
}
