% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization.R
\name{mstPlot}
\alias{mstPlot}
\title{mstPlot}
\usage{
mstPlot(mst = NULL, reduction = NULL, color.by = NULL,
  line.size = 0.5, line.color = "gray")
}
\arguments{
\item{mst}{mst object}

\item{reduction}{reduction array}

\item{color.by}{color array}

\item{line.size}{width of cell trajectory}

\item{line.color}{color of cell trajectory}
}
\description{
plot cells with MST
}
