% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{objectiveFunctionCpp}
\alias{objectiveFunctionCpp}
\title{objective function for gradient descnet method}
\usage{
objectiveFunctionCpp(alpha, X, Y)
}
\arguments{
\item{alpha}{current value of parameter alpha.}

\item{X}{distance vector.}

\item{Y}{distance vector.}
}
\description{
objective function for gradient descnet method
}
