% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildGraphTree.R
\name{orderCellsMST}
\alias{orderCellsMST}
\title{orderCellsMST}
\usage{
orderCellsMST(data = NULL, root = NULL, cutoff = 0.99)
}
\arguments{
\item{data}{a MST}

\item{root}{user choose the a root node}

\item{cutoff}{setup a percentile cutoff for scaling distance to pseudoTime, can avoid some outliers}
}
\description{
calculate pseduotime using a given start point from MST
}
