% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization.R
\name{plotPseudoTimeBox}
\alias{plotPseudoTimeBox}
\title{plotPseudoTimeBox}
\usage{
plotPseudoTimeBox(object = NULL, pseudotime = NULL,
  group.by = "ident", show.points = TRUE, fig.type = "box",
  coord.flip = TRUE, point.size = 0.2)
}
\arguments{
\item{object}{Seurat object}

\item{pseudotime}{name of pseudotime}

\item{group.by}{cell clusters}

\item{show.points}{display jetter points}

\item{fig.type}{show "box" plot or "vln" plot}

\item{coord.flip}{flip x and y axis}

\item{point.size}{size of jetter points}
}
\description{
plot pseudotime of each cluster on a Box or vln plot
}
