% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization.R
\name{plotPseudoTimeRidge}
\alias{plotPseudoTimeRidge}
\title{plotPseudoTimeRidge}
\usage{
plotPseudoTimeRidge(object = NULL, pseudotime = NULL,
  group.by = "ident", coord.flip = FALSE)
}
\arguments{
\item{object}{Seurat object}

\item{pseudotime}{name of pseudotime}

\item{group.by}{cell clusters}

\item{coord.flip}{flip x and y axis}
}
\description{
plot pseudotime of each cluster on a Ridge plot
}
