% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization.R
\name{plotRiverStream}
\alias{plotRiverStream}
\title{plotRiverStream.}
\usage{
plotRiverStream(object = NULL, pseudotime = NULL, group.by = "ident",
  display.legend = TRUE, legend.col = 1, colors = "auto",
  rel_widths = c(6, 1))
}
\value{
base plot streamgraph figure
}
\description{
plotRiverStream.
}
