% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization.R
\name{pseudoTimeMSTPlot}
\alias{pseudoTimeMSTPlot}
\title{pseudoTimeMSTPlot}
\usage{
pseudoTimeMSTPlot(object = NULL, reduction = NULL, pg = NULL,
  pseudotime = NULL, colors = c("blue", "green", "yellow", "red"),
  line.size = 0.5, line.color = "gray")
}
\arguments{
\item{object}{seurat object}

\item{reduction}{reduction name}

\item{pg}{principle graph (PG) used}

\item{pseudotime}{pseudotime used}

\item{colors}{gradient colors for pseudotime. color order is low to high.}

\item{line.size}{line width for trajectory}

\item{line.color}{line color for trajectory}
}
\description{
plot cells with trajectory and pseudoTime (MST based)
}
