% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization.R
\name{pseudoTimePlotKNN}
\alias{pseudoTimePlotKNN}
\title{pseudoTimePlotKNN}
\usage{
pseudoTimePlotKNN(object = NULL, assay = "Joint", colors = c("blue",
  "green", "yellow", "red"), line.size = 0.5, line.color = "gray")
}
\arguments{
\item{object}{seurat object}

\item{assay}{Assay name}

\item{colors}{gradient colors for pseudotime. color order is low to high.}

\item{line.size}{line width for kNN network}

\item{line.color}{line color for kNN network}
}
\description{
plot cells with trajectory and pseudoTime
}
