% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization.R
\name{pseudoTimePlotMST}
\alias{pseudoTimePlotMST}
\title{pseudoTimePlotMST}
\usage{
pseudoTimePlotMST(object = NULL, assay = NULL,
  show.trajectory = TRUE, colors = c("blue", "green", "yellow", "red"),
  line.size = 0.5, line.color = "gray")
}
\arguments{
\item{object}{seurat object}

\item{assay}{assay name (RNA, ADT or Joint)}

\item{show.trajectory}{show trajectory or not}

\item{colors}{gradient colors for pseudotime. color order is low to high.}

\item{line.size}{line width for trajectory}

\item{line.color}{line color for trajectory}
}
\description{
plot cells with trajectory and pseudoTime (MST based)
}
