% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImportData.R
\name{readDataFrom10X}
\alias{readDataFrom10X}
\title{readDataFrom10X}
\usage{
readDataFrom10X(dir = NULL, gene.column = 2, unique.features = TRUE)
}
\arguments{
\item{dir}{path of 10X cell ranger output folder}

\item{gene.column}{Specify which column of genes.tsv or features.tsv to use for gene names; default is 2}

\item{unique.features}{Make feature names unique (default TRUE)}
}
\description{
read data from 10X (cell ranger V3 output) by calling Seurat Read10X function
}
