% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PreProcess.R
\name{removeGene}
\alias{removeGene}
\title{removeGene}
\usage{
removeGene(object = NULL, pattern = NULL)
}
\arguments{
\item{object}{Seurat object}

\item{pattern}{the pattern (regular expression) of unwanted genes (e.g. '^IG[HKL]' for IgH, IgK, IgL genes). For more information about regular expression, please visit https://en.wikipedia.org/wiki/Regular_expression}
}
\description{
remove unwanted genes by gene name pattern (e.g. Ig genes)
For some cells, the expression of some specific genes may not related to the real cell popultion but will affect the cell clustering results. User can remove those genes from raw data before perform further analysis
An example is Ig genes for B cells.
}
