% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PreProcess.R
\name{removeProtein}
\alias{removeProtein}
\title{removeProtein}
\usage{
removeProtein(object = NULL, names = NULL)
}
\arguments{
\item{object}{Seurat object}

\item{names}{the name of proteins you want to remove from current analysis}
}
\description{
remove unwanted surface protein signals by protein names
}
