% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{scaleDistCpp}
\alias{scaleDistCpp}
\title{Scale distance using a sigmoid function. The main purpose is to reduce the effects of small distances (those distances are most likely from random noise of ADT signals)}
\usage{
scaleDistCpp(data, n, k)
}
\arguments{
\item{data}{data matrix of ADT.}

\item{n}{n for Sigmoid function.}

\item{k}{k for Sigmoid function.}
}
\description{
Scale distance using a sigmoid function. The main purpose is to reduce the effects of small distances (those distances are most likely from random noise of ADT signals)
}
