% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{scaleDistUpdateCpp}
\alias{scaleDistUpdateCpp}
\title{Scale distance using a sigmoid function. The main purpose is to reduce the effects of small distances (those distances are most likely from random noise of ADT signals)
This is an updated function. This function will distingrish distances between negative and positive from those within negative/positive groups. Only distances within negative/positive groups will be scaled.
This function may take longer time than normal one...}
\usage{
scaleDistUpdateCpp(data, n, k, c)
}
\arguments{
\item{data}{data matrix of ADT.}

\item{n}{n for Sigmoid function.}

\item{k}{k for Sigmoid function.}

\item{c}{c is a vector that contains constant value that seaprate negative and positive for each ADT feature. By default, c = 1,1,...,1}
}
\description{
Scale distance using a sigmoid function. The main purpose is to reduce the effects of small distances (those distances are most likely from random noise of ADT signals)
This is an updated function. This function will distingrish distances between negative and positive from those within negative/positive groups. Only distances within negative/positive groups will be scaled.
This function may take longer time than normal one...
}
