% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JointAnalysis.R
\name{scoreADT}
\alias{scoreADT}
\title{scoreADTfun}
\usage{
scoreADT(object = NULL, group.by = NULL, k = "auto",
  samples = NULL)
}
\arguments{
\item{object}{single cell data object}

\item{group.by}{name of clustering}

\item{k}{parameter k option. set "auto" will use the sum of SDs of entire dataset}

\item{samples}{name of sample information, can be blank}
}
\description{
ADT score function
}
