% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JointAnalysis.R
\name{scoreADTfun}
\alias{scoreADTfun}
\title{scoreADTfun}
\usage{
scoreADTfun(object = NULL, labels = NULL, labels.name = NULL,
  k = 10)
}
\arguments{
\item{object}{single cell data object}

\item{labels}{name of clustering}

\item{labels.name}{a list of all labels}

\item{k}{parameter k}
}
\description{
ADT score function
}
