% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JointAnalysis.R
\name{scoreRNA}
\alias{scoreRNA}
\title{scoreRNA}
\usage{
scoreRNA(object = NULL, group.by = NULL, platform = "UMI",
  span = 0.6, samples = NULL)
}
\arguments{
\item{object}{single cell data object}

\item{group.by}{name of clustering}

\item{platform}{platform of rna sequencing}

\item{span}{parameter span}

\item{samples}{name of sample information, can be blank}
}
\description{
RNA score function. use ROGUE score for RNA data
}
