% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Sigmoid}
\alias{Sigmoid}
\title{Sigmoid function}
\usage{
Sigmoid(a, n, k)
}
\arguments{
\item{a}{A Numeric Vector. original distance}

\item{n}{n for Sigmoid function.}

\item{k}{k for Sigmoid function.}
}
\description{
Sigmoid function
}
