% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PreProcess.R, R/generics.R
\name{softThreshold.Seurat}
\alias{softThreshold.Seurat}
\alias{softThreshold.default}
\alias{softThreshold}
\title{softThreshold.Seurat}
\usage{
\method{softThreshold}{Seurat}(object = NULL, sealing = 10)

\method{softThreshold}{default}(object = NULL, sealing = 10)

softThreshold(object, ...)
}
\arguments{
\item{object}{(For Seurat) Seurat object}

\item{sealing}{(For default) the sealing point of mitochondrial gene threshold, for 10X data we suggest use 10 percent}

\item{data}{(For default) Array of mito percent}
}
\description{
softThreshold
}
\details{
determine the soft threshold of percentage of mitochondrial gene. The criteria is: for each sample, we only keep the 95 percent cells with lowest percent of mito genes. If the cell quality is too bad, we use 10 percent as a sealing threshold
}
