% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization.R
\name{streamGraph}
\alias{streamGraph}
\title{Plots the streamgraph. This function was adopted from https://canonicalized.com/streamgraphs-in-tableau-via-r/ with modification, author is Dorian Banutoiu}
\usage{
streamGraph(yy, cols, plotTitle = "Streamgraph", streamNames = c())
}
\value{
base plot streamgraph figure
}
\description{
Plots the streamgraph. This function was adopted from https://canonicalized.com/streamgraphs-in-tableau-via-r/ with modification, author is Dorian Banutoiu
}
