% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization.R
\name{trajectoryKNNPlot}
\alias{trajectoryKNNPlot}
\title{trajectoryKNNPlot}
\usage{
trajectoryKNNPlot(object = NULL, group.by = NULL, reduction = NULL,
  graph = NULL, line.size = 0.5, line.color = "gray")
}
\arguments{
\item{object}{seurat object}

\item{group.by}{cell clusters}

\item{reduction}{reduction name}

\item{graph}{knn graph used}

\item{line.size}{width of cell trajectory}

\item{line.color}{color of cell trajectory}
}
\description{
plot cells with trajectory
}
