% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization.R
\name{trajectoryPlotMST}
\alias{trajectoryPlotMST}
\title{trajectoryPlotMST}
\usage{
trajectoryPlotMST(object = NULL, assay = "Joint", line.size = 0.5,
  line.color = "gray")
}
\arguments{
\item{object}{seurat object}

\item{assay}{Assay name}

\item{line.size}{width of cell trajectory}

\item{line.color}{color of cell trajectory}
}
\description{
plot cells with trajectory (MST based)
}
