% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildGraphTree.R
\name{trimMST}
\alias{trimMST}
\title{trimMST}
\usage{
trimMST(mst = NULL, embedding = NULL)
}
\arguments{
\item{mst}{a MST}

\item{embedding}{cell embedding}
}
\description{
trim MST, remove leaf. This step create initial tree for principle graph method.
}
