% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildGraphTree.R
\name{trimMSTcluster}
\alias{trimMSTcluster}
\title{trimMSTcluster}
\usage{
trimMSTcluster(mst = NULL, embedding = NULL, cluster.label = NULL,
  dist = NULL, nodes.per.cluster = 5)
}
\arguments{
\item{mst}{a MST}

\item{embedding}{cell embedding}

\item{cluster.label}{clustering label in object-meta}

\item{dist}{distance matrix for current assay}

\item{nodes.per.cluster}{how many nodes picked from each cluster}
}
\description{
generate a mini MST using cluster nodes as initial for PG
}
