% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JointAnalysis.R, R/generics.R
\name{tsneFromDistane.Seurat}
\alias{tsneFromDistane.Seurat}
\alias{tsneFromDistane.default}
\alias{tsneFromDistane}
\title{tsneFromDistane.Seurat}
\usage{
\method{tsneFromDistane}{Seurat}(object, assay = "All",
  perplexity = 30, dim = 2, seed = 42, theta = 0.5)

\method{tsneFromDistane}{default}(object, perplexity = 30, dim = 2,
  seed = 42, theta = 0.5)

tsneFromDistane(object, ...)
}
\arguments{
\item{object}{(For Seurat) Seurat object}

\item{assay}{(For Seurat) run t-SNE for which assay, choose from RNA, ADT, Joint or All}

\item{perplexity}{numeric; Perplexity parameter (should not be bigger than 3 * perplexity < nrow(X) - 1, see details for interpretation)}

\item{dim}{integer; Output dimensionality (default: 2)}

\item{seed}{integer; seed for reproducible results.}

\item{theta}{numeric; Speed/accuracy trade-off (increase for less accuracy), set to 0.0 for exact TSNE (default: 0.5)}
}
\description{
Run t-SNE to project single cells into 2D map using cell-cell pairwise distances
}
