% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JointAnalysis.R, R/generics.R
\name{umapFromDistane.Seurat}
\alias{umapFromDistane.Seurat}
\alias{umapFromDistane.default}
\alias{umapFromDistane}
\title{umapFromDistane.Seurat}
\usage{
\method{umapFromDistane}{Seurat}(object, assay = "All", seed = 42,
  method = "umap-learn", n.neighbors = 15, n.components = 2,
  metric = "euclidean", verbose = FALSE, n.epochs = 200,
  min.dist = 0.1, spread = 1, set.op.mix.ratio = 1,
  local.connectivity = 1L, negative.sample.rate = 5L)

\method{umapFromDistane}{default}(object, seed = 42,
  method = "umap-learn", n.neighbors = 15, n.components = 2,
  metric = "euclidean", verbose = TRUE, n.epochs = 200,
  min.dist = 0.1, spread = 1, set.op.mix.ratio = 1,
  local.connectivity = 1L, negative.sample.rate = 5L)

umapFromDistane(object, ...)
}
\arguments{
\item{object}{(For Seurat) Seurat object}

\item{assay}{(For Seurat) run UMAP for which assay, choose from RNA, ADT, Joint or All}

\item{seed}{see number. default is 42}

\item{method}{could be "naive" or "umap-learn"}

\item{n.neighbors}{integer; number of nearest neighbors}

\item{n.components}{integer; dimension of target (output) space}

\item{metric}{character or function; determines how distances between data points are computed. When using a string, available metrics are: euclidean, manhattan. Other available generalized metrics are: cosine, pearson, pearson2. Note the triangle inequality may not be satisfied by some generalized metrics, hence knn search may not be optimal. When using metric.function as a function, the signature must be function(matrix, origin, target) and should compute a distance between the origin column and the target columns}

\item{verbose}{logical or integer; determines whether to show progress messages}

\item{n.epochs}{integer; number of iterations performed during layout optimization}

\item{min.dist}{numeric; determines how close points appear in the final layout}

\item{spread}{numeric; used during automatic estimation of a/b parameters.}

\item{set.op.mix.ratio}{numeric in range [0,1]; determines who the knn-graph is used to create a fuzzy simplicial graph}

\item{local.connectivity}{numeric; used during construction of fuzzy simplicial set}

\item{negative.sample.rate}{integer; determines how many non-neighbor points are used per point and per iteration during layout optimization}
}
\description{
Run UMAP to project single cells into 2D space using cell-cell pairwise distances
}
