% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization.R
\name{generateGridDimPlot}
\alias{generateGridDimPlot}
\title{generateGridDimPlot}
\usage{
generateGridDimPlot(object = NULL, assays = c("RNA", "ADT", "Joint"),
  darkTheme = TRUE, legend = TRUE, cluster.label = TRUE,
  cluster.lable.size = 5, reduction = "tsne")
}
\arguments{
\item{object}{seurat object}

\item{assays}{assays user want to plot and compare}

\item{darkTheme}{switch for darkTheme (TRUE or FALSE)}

\item{legend}{switch for showing legend in each sub plot (TRUE or FALSE)}

\item{cluster.label}{swicth for showing cluster label on each sub plot (TRUE or FALSE)}

\item{cluster.lable.size}{font size for cluster labels}

\item{reduction}{reduction for all assays. Choose from tsne or umap}
}
\description{
generate a grid dim plot object list for RNA, ADT and joint analysis
}
