% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JointAnalysis.R, R/generics.R
\name{jointDistance.Seurat}
\alias{jointDistance.Seurat}
\alias{jointDistance.default}
\alias{jointDistance}
\title{jointDistance.Seurat}
\usage{
\method{jointDistance}{Seurat}(
  object,
  dims = 20,
  beta = 0.5,
  model = "LP",
  keep.rna = TRUE,
  keep.adt = TRUE,
  sigmoid.n = 10,
  sigmoid.k = 0.5,
  precision.mode = TRUE,
  c = NULL
)

\method{jointDistance}{default}(object, dist1 = NULL, dist2 = NULL, beta = 1, model = "LP")

jointDistance(object, ...)
}
\arguments{
\item{object}{(For Seurat) Seurat object}

\item{dims}{(For Seurat) number of PCs used for RNA data. Default is 20}

\item{beta}{(For default) For L1 model, use beta to balence contributions from RNA and ADT in the joint distances (0 <= beta <= 1). Default value is 0.5, indicate L1 modal will treat RNA and ADT equally. higher beta will increase the contribution of RNA}

\item{model}{(For default) Norm model for distance integration. Default model is L-infinite model. Another option is L1 model.}

\item{dist1}{(For default) dist object of first assay}

\item{dist2}{(For default) dist object of second assay}
}
\description{
If input seurat object, will calculate cell-cell pairwise distances for RNA and ADT separately, then calculate the joint cell-cell pairwise distances. It alao can calculate joint distances for any two dist objects
}
