% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildGraphTree.R
\name{orderCellsKNN}
\alias{orderCellsKNN}
\title{orderCellsKNN}
\usage{
orderCellsKNN(data = NULL, dist = NULL, root = NULL)
}
\arguments{
\item{data}{a KNN}

\item{dist}{dist matrix the KNN calculated from}

\item{root}{user choose the a root node}
}
\description{
calculate pseduotime using a given start point from KNN
}
