% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization.R
\name{plotRiverStream}
\alias{plotRiverStream}
\title{plotRiverStream.}
\usage{
plotRiverStream(object = NULL, pseudotime = NULL, group.by = "ident",
  display.legend = TRUE, legend.col = 1, colors = "auto",
  rel_widths = c(6, 1), display.arrow = TRUE)
}
\arguments{
\item{object}{seurat object}

\item{pseudotime}{name of pseudotime, e.g. jointMSTtime}

\item{group.by}{name of cell clusters. e.g. jointClusterID}

\item{display.legend}{display figure legend or not}

\item{legend.col}{display all legend elements in n columns, default = 1}

\item{colors}{array of color names. default is "auto"}

\item{rel_widths}{reletive width between river plot and figure legend. By default is c(6,1), indicates figure:legend = 6:1}

\item{display.arrow}{display arrow or not}
}
\description{
plotRiverStream.
}
