% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildGraphTree.R
\name{pseudoTime}
\alias{pseudoTime}
\title{pseudoTime}
\usage{
pseudoTime(object, method = c("MST", "KNN"), assay = "Joint",
  root.cluster = NULL)
}
\arguments{
\item{object}{Seurat object}

\item{method}{could be MST or KNN}

\item{assay}{Assay name. Can be RNA, ADT or Joint}

\item{root.cluster}{user choose the root cell cluster}
}
\description{
calculate pseudotime using a given start point from MST or KNN.
}
